Update import process
Now expecting the CSV from black pug, so we don't have to manually
generate the WeekXDataMaster.xlsx spreadsheet. We _don't_ get
information on what sites people are staying in, so that gets an extra
page for manual import for now. Hopefully, we'll be able to do that
automatically at some point, paging @alex.ferrell... :)

We no longer depend on PhpSpreadsheet.

We also fix a few bugs where we pass a `Week` object, but that gets
dependency injected, so we just take the ID and `Week::find()` it
ourselves.

We also now try to do all our grouping by unit and by council, rather
than just by unit; it's not uncommon to have, e.g. this week 104 Mid
Iowa Council and 104 Mid America Council, so we need to be able to
differentiate those.