Add script to import data to multi-week setup
At the beginning of the year, we were restting the database every week.
We now have a multi-week model, but we'd like to include all the data
from previous weeks before we had the multi-week setup working. This
script takes a single-week database and imports it into the existing
multi-week database.